Moved cleanup out of background thread to stick with the architecture… (#4823)
* Moved cleanup out of background thread to stick with the architecture tests

* remove unused imports

* Had to split the URLUtil since another logic class used URLUtil functionality

* fixed checkstyle

* Update DOAJFetcher.java

* checkstyle III
