+++ src/main/java/org/jabref/gui/actions/CleanupAction.java
@@ -9,7 +9,6 @@ import org.jabref.gui.DialogService;
 import org.jabref.gui.cleanup.CleanupDialog;
 import org.jabref.gui.undo.NamedCompound;
 import org.jabref.gui.undo.UndoableFieldChange;
-import org.jabref.gui.util.BackgroundTask;
 import org.jabref.logic.cleanup.CleanupPreset;
 import org.jabref.logic.cleanup.CleanupWorker;
 import org.jabref.logic.l10n.Localization;
@@ -59,9 +58,8 @@ public class CleanupAction implements BaseAction {
 
             preferences.setCleanupPreset(chosenPreset.get());
 
-            BackgroundTask.wrap(() -> cleanup(chosenPreset.get()))
-                          .onSuccess(result -> showResults())
-                          .executeWith(Globals.TASK_EXECUTOR);
+            this.cleanup(chosenPreset.get());
+            this.showResults();
         }
     }
